refactor(support): move the flow jest config out of features/platform
@features/platform-jest-config was development-only tooling — a jest config
factory, a jsdom setup file and generic Lumen passthrough stubs — sitting in
features/platform and carrying scope:features + type:feature-platform tags. It
becomes @support/jest-features-flow, matching the product preset the ADR names
for features/flow/*.

Its API is unchanged; the three features/flow consumers only swap the package
name. Its tsconfig.json turns out to have been dead: moduleResolution node10 was
removed in the TypeScript version this repo uses, so it could not have been
running, and the package had no typecheck script to run it. It now inherits
resolution from the base config, with typecheck and format scripts added.
