+++ CODEOWNERS
@@ -77,7 +77,8 @@ devtools/wire/                                           @ledgerhq/platform
 devtools/relay/                                          @ledgerhq/platform
 
 # Platform — support
-support/jest-devtools/                           @ledgerhq/platform
+support/jest-devtools/                                   @ledgerhq/platform
+support/jest-features-flow/                          @ledgerhq/platform
 
 # Wallet — libs
 libs/live-countervalues-react/                           @ledgerhq/wallet-xp
