+++ support/jest-features-flow/mocks/file-stub.js
