+++ support/jest-features-flow/mocks/react-native.js
