+++ support/jest-features-flow/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@features/platform-jest-config",
+  "name": "@support/jest-features-flow",
   "version": "0.1.0",
   "private": true,
   "description": "Shared jest config + generic Lumen stubs for features/flow packages (dual web/native projects)",
@@ -14,6 +14,11 @@
     "mocks",
     "setup"
   ],
+  "scripts": {
+    "format": "oxfmt index.js mocks setup",
+    "format:check": "oxfmt index.js mocks setup --check",
+    "typecheck": "tsc --noEmit"
+  },
   "peerDependencies": {
     "react": ">=19.0.0"
   },
