+++ support/jest-features-flow/tsconfig.json
@@ -1,10 +1,8 @@
 {
-  "extends": "../../../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "target": "ES2022",
     "lib": ["ES2022"],
-    "module": "CommonJS",
-    "moduleResolution": "node",
     "allowJs": true,
     "checkJs": false,
     "noEmit": true
